requirements.txt
# Flask framework
Flask==3.1.0
Jinja2==3.1.5
Flask-WTF==1.2.2
Flask-Session==0.8.0
Flask-Mail==0.9.1

# HTTP requests (for calling ML API)
requests==2.32.3

# Image handling (for uploads)
Pillow==10.4.0

# Database
mysql-connector-python==9.0.0

# Environment management
python-dotenv==1.0.1

# Security and authentication
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# OAuth authentication
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Google authentication
google-auth==2.25.2
google-auth-oauthlib==1.1.0

# Gunicorn for production
gunicorn==21.2.0
